+++ src/carp/oob_buffer.h
@@ -33,7 +33,7 @@ namespace carp {
 typedef struct particle_mem {
   float indexed_prop;             // float key for range q (cfg via CarpOptions)
   char buf[CARP_MAXPARTSZ];       // buf w/encoded particle (key+filename+data)
-  int buf_sz;                     // total size of encoded data in buf[]
+  size_t buf_sz;                  // total size of encoded data in buf[]
   int shuffle_dest;               // rank# or -1 (unk), via AssignShuffleTarget
 } particle_mem_t;
